mindmap: upload the text version
Note that reading those notes by pure text is not recommended since
some of them are images and not be able to be exported as pure text.

Also those maps are constructed by a certain relation/connection, which
I doubt that the exported pure text can represent the same meaning.